Add crud functionalities to the stock microservice

+++ src/main/java/com/pragma/powerup/stockservice/domains/model/Brand.java
@@ -1,24 +1,24 @@
 package com.pragma.powerup.stockservice.domains.model;
 
 public class Brand {
-    private Integer id;
+    private Long id;
     private String name;
     private String description;
 
     public Brand() {
     }
 
-    public Brand(Integer id, String name, String description) {
+    public Brand(Long id, String name, String description) {
         this.id = id;
         this.name = name;
         this.description = description;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
